fraction.cpp

+++ fraction.cpp
@@ -5,9 +5,9 @@ struct Fraction {
     int denominator;
 };
 Fraction& largestFraction(Fraction&f1,Fraction&f2,Fraction&f3){
-    double value1=f1.numerator/f1.denominator;
-    double value2=f2.numerator/f2.denominator;
-    double value3=f3.numerator/f3.denominator;
+    double value1=(double)f1.numerator/f1.denominator;
+    double value2=(double)f2.numerator/f2.denominator;
+    double value3=(double)f3.numerator/f3.denominator;
 if (value1>=value2 and value1>=value3)
   return f1;
 else if (value2>=value1 and value2>=value3)
